+++ README.md
@@ -1,0 +1,65 @@
+Консольное приложение для Clever-Bank.
+
+Описание: консольное приложение для сотрудников банка, в котором можно осуществлять операции по обслуживанию Клиентов банка.
+
+Запуск приложения: 
+1. запустить docker для работы с базой данных 
+   в командной строке в папке, где находится файл docker-compose.yml запустить команду docker-compose up -d
+   ![img.png](img/img1.png)
+2. запустить в классе App метод main().
+
+После запуска приложения должны автоматически создаться таблицы в БД и заполниться данными.
+   ![img.png](img/img2.png)
+   ![img.png](img/img3.png)
+Также в консоли выведется Главное Меню для работы.
+   ![img.png](img/img4.png)
+При вводе 1 переходим в Меня операций над Банками.
+   ![img.png](img/img5.png)
+При вводе 2 переходим в Меня операций над Клиентами.
+   ![img.png](img/img6.png)
+При вводе 3 переходим в Меня операций над Счетами.
+   ![img.png](img/img7.png)
+
+Данное приложение работает с консолью, поэтому необходимо выбирать пункт того, что хотим сделать и дальше вводить в консоль, запрашиваемую приложением информацию.
+
+Например, мы хотим добавить новый банк. Войдем в Меня операций над банками, выберем пункт "добавление банка" и введем наименование банка. Новый банк создан.
+   ![img.png](img/img8.png)
+   ![img.png](img/img9.png)
+Создадим нового Клиента. Войдем в Меня операций над клиентами, выберем пункт "добавление клиента" и введем Имя, Отчество, Фамилию и Адрес Клиента. Новый клиент создан.
+   ![img.png](img/img10.png)
+   ![img.png](img/img11.png)
+После создания Клиента необходимо его добавить в Банк.
+   ![img.png](img/img12.png)
+   ![img.png](img/img13.png)
+Создадим новый счет. Войдем в Меню операций над Счетами. Введем пункт "добавление счета", введем id клиента и id банка и счет открыт.
+   ![img.png](img/img14.png)
+Пополним данный счет на 5000 BYN.
+   ![img.png](img/img15.png)
+Также будет сформирован Банковский счет на данную операцию.
+   ![img.png](img/img16.png)
+Переведем 1000 BYN Михайловне Александре Никитичне.
+   ![img.png](img/img17.png)
+   ![img.png](img/img18.png)
+Сделаем перевод денег от Богданова Михаила Яковлевича - на счет Иванова Ивана Ивановича в размере 500 BYN.
+   ![img.png](img/img19.png)
+   ![img.png](img/img20.png)
+Снимем 2000 BYN со счет Иванова Ивана Ивановича.
+   ![img.png](img/img21.png)
+   ![img.png](img/img22.png)
+Автоматически по расписанию в последний день месяца будет начислен 1% дохода на остаток денежных средств на счету. В образце (для отображения данного функционала проценты были начислены 02.09).
+   ![img.png](img/img23.png)
+Сформируем выписку по счету Иванова Ивана Ивановича за 02.09.2023
+   ![img.png](img/img24.png)
+
+При проведении банковских операций происходит проверка существует ли данный Счет, наличия достаточного количество денег для операции и не ввел ли пользователь отрицательное число.
+   ![img.png](img/img25.png)
+   ![img.png](img/img26.png)
+   ![img.png](img/img27.png)
+
+Таким образом данное консольное приложение позволяет выполнять CRUD операции над Банками, Клиентами и Счетами.
+Позволяет осуществлять Пополнение Счета, Перевод Денег на другие Счета, Снятие денег, а также автоматическое начисление процентов раз в месяц на остаток денежных средств на счету.
+После каждой операции происходит формирование подтверждающего документа об операции.
+Также имеется возможность формировать выписку по счету об всех операциях за определенный период времени.
+
+
+
